# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /user:
    get:
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户列表查询
      tags:
      - 用户模块
  /user/changepwd:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 旧密码
        in: formData
        name: oldPassword
        type: string
      - description: 新密码
        in: formData
        name: newPassword
        type: string
      - description: 确认新密码
        in: formData
        name: checkNewPassword
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改密码
      tags:
      - 用户模块
  /user/create:
    get:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 手机号码
        in: query
        name: phone
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户模块
  /user/delete:
    get:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/login:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 登录密码
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户登录
      tags:
      - 用户模块
  /user/updata:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改用户手机和邮箱
      tags:
      - 用户模块
swagger: "2.0"